# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Lamsa - Beauty & Gifts Delivery API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: products
    description: Product operations
  - name: categories
    description: Category operations
  - name: stores
    description: Store operations
  - name: cart
    description: Cart operations
  - name: orders
    description: Order operations
  - name: reviews
    description: Review operations
  - name: promotions
    description: Promotions and coupons
  - name: dashboard
    description: Dashboard and stats

paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /categories:
    get:
      operationId: listCategories
      tags: [categories]
      summary: List all categories
      responses:
        "200":
          description: List of categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"

  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List products with optional filters
      parameters:
        - name: categoryId
          in: query
          schema:
            type: ["integer", "null"]
        - name: storeId
          in: query
          schema:
            type: ["integer", "null"]
        - name: search
          in: query
          schema:
            type: ["string", "null"]
        - name: featured
          in: query
          schema:
            type: ["boolean", "null"]
        - name: limit
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: List of products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
    post:
      operationId: createProduct
      tags: [products]
      summary: Create a new product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductInput"
      responses:
        "201":
          description: Created product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"

  /products/{id}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get product by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Product details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          description: Not found

  /products/featured:
    get:
      operationId: getFeaturedProducts
      tags: [products]
      summary: Get featured/best-selling products
      responses:
        "200":
          description: Featured products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"

  /products/new-arrivals:
    get:
      operationId: getNewArrivals
      tags: [products]
      summary: Get newly added products
      responses:
        "200":
          description: New arrival products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"

  /products/top-sanaa:
    get:
      operationId: getTopSanaaProducts
      tags: [products]
      summary: Get most ordered products in Sanaa
      responses:
        "200":
          description: Top products in Sanaa
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"

  /stores:
    get:
      operationId: listStores
      tags: [stores]
      summary: List all stores
      responses:
        "200":
          description: List of stores
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Store"
    post:
      operationId: createStore
      tags: [stores]
      summary: Create a new store
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StoreInput"
      responses:
        "201":
          description: Created store
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Store"

  /stores/{id}:
    get:
      operationId: getStore
      tags: [stores]
      summary: Get store by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Store details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Store"

  /stores/featured:
    get:
      operationId: getFeaturedStores
      tags: [stores]
      summary: Get featured stores
      responses:
        "200":
          description: Featured stores
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Store"

  /cart:
    get:
      operationId: getCart
      tags: [cart]
      summary: Get current cart
      parameters:
        - name: sessionId
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Cart details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"
    post:
      operationId: addToCart
      tags: [cart]
      summary: Add item to cart
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CartItemInput"
      responses:
        "200":
          description: Updated cart
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"

  /cart/{itemId}:
    patch:
      operationId: updateCartItem
      tags: [cart]
      summary: Update cart item quantity
      parameters:
        - name: itemId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CartItemUpdate"
      responses:
        "200":
          description: Updated cart
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"
    delete:
      operationId: removeCartItem
      tags: [cart]
      summary: Remove item from cart
      parameters:
        - name: itemId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated cart
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"

  /orders:
    get:
      operationId: listOrders
      tags: [orders]
      summary: List orders for a user session
      parameters:
        - name: sessionId
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
    post:
      operationId: createOrder
      tags: [orders]
      summary: Place a new order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderInput"
      responses:
        "201":
          description: Created order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"

  /orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get order details by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Order details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"

  /reviews:
    get:
      operationId: listReviews
      tags: [reviews]
      summary: List reviews for a product
      parameters:
        - name: productId
          in: query
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of reviews
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Review"
    post:
      operationId: createReview
      tags: [reviews]
      summary: Submit a review
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReviewInput"
      responses:
        "201":
          description: Created review
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Review"

  /promotions/banners:
    get:
      operationId: listBanners
      tags: [promotions]
      summary: Get promotional banners
      responses:
        "200":
          description: Promotional banners
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Banner"

  /promotions/coupons/validate:
    post:
      operationId: validateCoupon
      tags: [promotions]
      summary: Validate a coupon code
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CouponValidateInput"
      responses:
        "200":
          description: Coupon validation result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CouponValidateResult"

  /dashboard/stats:
    get:
      operationId: getDashboardStats
      tags: [dashboard]
      summary: Get dashboard summary statistics
      responses:
        "200":
          description: Dashboard statistics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardStats"

components:
  schemas:
    HealthStatus:
      type: object
      required: [status]
      properties:
        status:
          type: string

    Category:
      type: object
      required: [id, nameAr, nameEn, icon, productCount]
      properties:
        id:
          type: integer
        nameAr:
          type: string
        nameEn:
          type: string
        icon:
          type: string
        color:
          type: ["string", "null"]
        productCount:
          type: integer
        imageUrl:
          type: ["string", "null"]

    Product:
      type: object
      required: [id, nameAr, price, categoryId, storeId, imageUrl, inStock, rating, reviewCount]
      properties:
        id:
          type: integer
        nameAr:
          type: string
        nameEn:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        price:
          type: number
        discountPrice:
          type: ["number", "null"]
        categoryId:
          type: integer
        categoryName:
          type: ["string", "null"]
        storeId:
          type: integer
        storeName:
          type: ["string", "null"]
        imageUrl:
          type: string
        images:
          type: array
          items:
            type: string
        brand:
          type: ["string", "null"]
        inStock:
          type: boolean
        stockQuantity:
          type: ["integer", "null"]
        rating:
          type: number
        reviewCount:
          type: integer
        isFeatured:
          type: boolean
        isNew:
          type: boolean
        deliveryDays:
          type: ["integer", "null"]
        tags:
          type: array
          items:
            type: string

    ProductInput:
      type: object
      required: [nameAr, price, categoryId, storeId, imageUrl]
      properties:
        nameAr:
          type: string
        nameEn:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        price:
          type: number
        discountPrice:
          type: ["number", "null"]
        categoryId:
          type: integer
        storeId:
          type: integer
        imageUrl:
          type: string
        brand:
          type: ["string", "null"]
        stockQuantity:
          type: ["integer", "null"]
        isFeatured:
          type: boolean
        deliveryDays:
          type: ["integer", "null"]

    Store:
      type: object
      required: [id, nameAr, logoUrl, rating, reviewCount, isOpen]
      properties:
        id:
          type: integer
        nameAr:
          type: string
        nameEn:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        logoUrl:
          type: string
        coverUrl:
          type: ["string", "null"]
        rating:
          type: number
        reviewCount:
          type: integer
        isOpen:
          type: boolean
        workingHours:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        location:
          type: ["string", "null"]
        isFeatured:
          type: boolean
        productCount:
          type: integer

    StoreInput:
      type: object
      required: [nameAr, logoUrl]
      properties:
        nameAr:
          type: string
        nameEn:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        logoUrl:
          type: string
        coverUrl:
          type: ["string", "null"]
        workingHours:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        location:
          type: ["string", "null"]

    Cart:
      type: object
      required: [sessionId, items, subtotal, deliveryFee, total]
      properties:
        sessionId:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/CartItem"
        subtotal:
          type: number
        deliveryFee:
          type: number
        discount:
          type: number
        total:
          type: number
        couponCode:
          type: ["string", "null"]

    CartItem:
      type: object
      required: [id, productId, productName, price, quantity, imageUrl]
      properties:
        id:
          type: integer
        productId:
          type: integer
        productName:
          type: string
        price:
          type: number
        quantity:
          type: integer
        imageUrl:
          type: string
        giftWrapping:
          type: boolean
        giftMessage:
          type: ["string", "null"]

    CartItemInput:
      type: object
      required: [sessionId, productId, quantity]
      properties:
        sessionId:
          type: string
        productId:
          type: integer
        quantity:
          type: integer
        giftWrapping:
          type: boolean
        giftMessage:
          type: ["string", "null"]

    CartItemUpdate:
      type: object
      required: [quantity]
      properties:
        quantity:
          type: integer

    Order:
      type: object
      required: [id, sessionId, status, total, createdAt, items]
      properties:
        id:
          type: integer
        sessionId:
          type: string
        status:
          type: string
          enum: [received, preparing, delivering, delivered]
        total:
          type: number
        deliveryFee:
          type: number
        discount:
          type: number
        deliveryAddress:
          type: ["string", "null"]
        customerName:
          type: ["string", "null"]
        customerPhone:
          type: ["string", "null"]
        paymentMethod:
          type: string
          enum: [cash_on_delivery, bank_transfer]
        couponCode:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"

    OrderItem:
      type: object
      required: [id, productId, productName, price, quantity, imageUrl]
      properties:
        id:
          type: integer
        productId:
          type: integer
        productName:
          type: string
        price:
          type: number
        quantity:
          type: integer
        imageUrl:
          type: string

    OrderInput:
      type: object
      required: [sessionId, deliveryAddress, customerName, customerPhone, paymentMethod]
      properties:
        sessionId:
          type: string
        deliveryAddress:
          type: string
        customerName:
          type: string
        customerPhone:
          type: string
        paymentMethod:
          type: string
          enum: [cash_on_delivery, bank_transfer]
        couponCode:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]

    Review:
      type: object
      required: [id, productId, rating, createdAt]
      properties:
        id:
          type: integer
        productId:
          type: integer
        customerName:
          type: string
        rating:
          type: integer
        comment:
          type: ["string", "null"]
        createdAt:
          type: string

    ReviewInput:
      type: object
      required: [productId, rating]
      properties:
        productId:
          type: integer
        customerName:
          type: string
        rating:
          type: integer
        comment:
          type: ["string", "null"]

    Banner:
      type: object
      required: [id, title, imageUrl]
      properties:
        id:
          type: integer
        title:
          type: string
        subtitle:
          type: ["string", "null"]
        imageUrl:
          type: string
        linkType:
          type: ["string", "null"]
        linkId:
          type: ["integer", "null"]
        isActive:
          type: boolean

    CouponValidateInput:
      type: object
      required: [code]
      properties:
        code:
          type: string
        cartTotal:
          type: ["number", "null"]

    CouponValidateResult:
      type: object
      required: [valid]
      properties:
        valid:
          type: boolean
        discountType:
          type: ["string", "null"]
        discountValue:
          type: ["number", "null"]
        discountAmount:
          type: ["number", "null"]
        message:
          type: ["string", "null"]

    DashboardStats:
      type: object
      required: [totalProducts, totalStores, totalOrders, totalCategories, featuredProductsCount, newArrivalsCount]
      properties:
        totalProducts:
          type: integer
        totalStores:
          type: integer
        totalOrders:
          type: integer
        totalCategories:
          type: integer
        featuredProductsCount:
          type: integer
        newArrivalsCount:
          type: integer
        topCategory:
          type: ["string", "null"]
        activePromotions:
          type: integer