# lib/api-zod/src/generated/types/banner.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Lamsa - Beauty & Gifts Delivery API
 * OpenAPI spec version: 0.1.0
 */

export interface Banner {
  id: number;
  title: string;
  /** @nullable */
  subtitle?: string | null;
  imageUrl: string;
  /** @nullable */
  linkType?: string | null;
  /** @nullable */
  linkId?: number | null;
  isActive?: boolean;
}