# lib/api-zod/src/generated/types/cartItem.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Lamsa - Beauty & Gifts Delivery API
 * OpenAPI spec version: 0.1.0
 */

export interface CartItem {
  id: number;
  productId: number;
  productName: string;
  price: number;
  quantity: number;
  imageUrl: string;
  giftWrapping?: boolean;
  /** @nullable */
  giftMessage?: string | null;
}