# lib/api-zod/src/generated/types/couponValidateResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Lamsa - Beauty & Gifts Delivery API
 * OpenAPI spec version: 0.1.0
 */

export interface CouponValidateResult {
  valid: boolean;
  /** @nullable */
  discountType?: string | null;
  /** @nullable */
  discountValue?: number | null;
  /** @nullable */
  discountAmount?: number | null;
  /** @nullable */
  message?: string | null;
}