# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Lamsa - Beauty & Gifts Delivery API
 * OpenAPI spec version: 0.1.0
 */
import type { OrderItem } from './orderItem';
import type { OrderPaymentMethod } from './orderPaymentMethod';
import type { OrderStatus } from './orderStatus';

export interface Order {
  id: number;
  sessionId: string;
  status: OrderStatus;
  total: number;
  deliveryFee?: number;
  discount?: number;
  /** @nullable */
  deliveryAddress?: string | null;
  /** @nullable */
  customerName?: string | null;
  /** @nullable */
  customerPhone?: string | null;
  paymentMethod?: OrderPaymentMethod;
  /** @nullable */
  couponCode?: string | null;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
  items: OrderItem[];
}